producer/requirements.txt
# Python dependencies for the Kafka producer
kafka-python>=2.0.2
